added login feature

+++ base/views.py
@@ -2,6 +2,11 @@ from django.shortcuts import render,redirect
 from django.db.models import Q
 from .models import Room,Topic
 from .forms import RoomForm
+from django.http import HttpResponse
+from django.contrib.auth.models import User
+from django.contrib.auth import authenticate,login,logout
+from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 # Create your views here.
 # rooms = [
 #     {'id':1,'name':'deluxe room'},
@@ -22,11 +27,13 @@ def home(request):
     context = {'rooms':rooms,'topics':topics,'room_count':room_count} 
     return render(request,'base/home.html',context)
 
+
 def room(request,pk):
     room = Room.objects.get(id=pk)
     context = {'room':room}
     return render(request,'base/room.html',context)
 
+@login_required(login_url='login')
 def createRoom(request):
     form = RoomForm()
     if request.method == 'POST':
@@ -37,22 +44,57 @@ def createRoom(request):
     context = {'form':form}
     return render(request,'base/room_form.html',context)
 
+@login_required(login_url='login')
 def updateRoom(request,pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
-    updated='hi'
+    
+    if request.user != room.host:
+        return HttpResponse('your are not authrized')
+
     if request.method=="POST":
         form = RoomForm(request.POST,instance=room)
         if form.is_valid():
             form.save()
             return redirect('home')
 
-    context = {'form':form,'updated':updated}
+    context = {'form':form}
     return render(request,'base/room_form.html',context)
 
+@login_required(login_url='login')
 def deleteRoom(request,pk):
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        return HttpResponse('your are not authrized')
+        
     if request.method=='POST':
         room.delete()
         return redirect('home')
     return render(request,'base/delete.html',{'obj':room})
+
+def loginPage(request):
+    if request.method=='POST':
+        username = request.POST.get('username')
+        password = request.POST.get('password')
+
+        try:
+            user = User.objects.get(username=username)
+        except:
+            messages.error(request,"user doesn't exists")
+
+        user = authenticate(request,username=username,password=password)
+
+        if user is not None:
+            login(request,user)
+            return redirect('home')
+        else:
+            messages.error(request,"Invalid credentials")
+
+
+    context={}
+    return render(request,'base/login_register.html',context)
+
+def logoutUser(request):
+    logout(request)
+    return redirect('home')
